+++ packages/vm/src/traits.rs
@@ -7,13 +7,13 @@ use cosmwasm_std::{Order, KV};
 /// Designed to allow easy dependency injection at runtime.
 /// This cannot be copied or cloned since it would behave differently
 /// for mock storages and a bridge storage in the VM.
-pub struct Extern<S: BackendStorage, A: Api, Q: Querier> {
+pub struct Extern<S: Storage, A: Api, Q: Querier> {
     pub storage: S,
     pub api: A,
     pub querier: Q,
 }
 
-impl<S: BackendStorage, A: Api, Q: Querier> Extern<S, A, Q> {
+impl<S: Storage, A: Api, Q: Querier> Extern<S, A, Q> {
     /// change_querier is a helper mainly for test code when swapping out the Querier
     /// from the auto-generated one from mock_dependencies. This changes the type of
     /// Extern so replaces requires some boilerplate.
@@ -54,7 +54,7 @@ impl<I: StorageIterator + ?Sized> StorageIterator for Box<I> {
 }
 
 /// Access to the VM's backend storage, i.e. the chain
-pub trait BackendStorage
+pub trait Storage
 where
     Self: 'static,
 {